+++ common/service-utils/package.json
@@ -11,8 +11,8 @@
     "build": "tsc -b"
   },
   "devDependencies": {
-    "@types/mongodb": "^3.5.25",
-    "@types/node": "^14.0.27"
+    "@types/mongodb": "^3.5.26",
+    "@types/node": "^14.6.0"
   },
   "dependencies": {
     "@common/config": "1.0.0",
