+++ common/service-utils/src/index.ts
@@ -1,5 +1,6 @@
 import './use-common-http-auth'
 import './use-db-logger'
+import './application-context'
 export * from './authorized-data-set'
 export * from './authorize-ownership'
 export * from './create-indexes'
