+++ common/service-utils/src/use-db-logger.ts
@@ -10,7 +10,6 @@ import '@furystack/repository'
 
 @Injectable({ lifetime: 'singleton' })
 export class DbLoggerSettings {
-  public appName!: string
   public minLevel?: LogLevel
 }
 
@@ -28,12 +27,13 @@ export class DbLogger extends AbstractLogger implements Disposable {
   }
   public readonly store: MongodbStore<diag.LogEntry<any>>
   public async addEntry<T>(entry: import('@furystack/logging').LeveledLogEntry<T>): Promise<void> {
+    const { name: appName } = this.injector.getApplicationContext()
     if (!this.isDisposing) {
       if (entry.level >= (this.settings.minLevel || LogLevel.Information)) {
         await this.store.add({
           ...entry,
           creationDate: new Date(),
-          appName: this.settings.appName,
+          appName,
         })
       }
     }
