+++ package.json
@@ -22,16 +22,16 @@
   },
   "optionalDependencies": {
     "@cypress/webpack-preprocessor": "^5.4.4",
-    "@typescript-eslint/eslint-plugin": "^3.9.0",
-    "@typescript-eslint/parser": "^3.9.0",
+    "@typescript-eslint/eslint-plugin": "^3.9.1",
+    "@typescript-eslint/parser": "^3.9.1",
     "concurrently": "^5.1.0",
     "cypress": "^4.12.1",
     "cypress-plugin-snapshots": "^1.4.4",
-    "eslint": "^7.6.0",
+    "eslint": "^7.7.0",
     "eslint-config-prettier": "^6.5.0",
     "eslint-plugin-cypress": "^2.7.0",
     "eslint-plugin-import": "^2.21.2",
-    "eslint-plugin-jsdoc": "^30.2.2",
+    "eslint-plugin-jsdoc": "^30.2.4",
     "eslint-plugin-prettier": "^3.1.4",
     "husky": "^4.0.6",
     "jest": "^26.4.0",
@@ -40,10 +40,10 @@
     "prettier": "^2.0.1"
   },
   "devDependencies": {
-    "@types/jest": "^26.0.9",
+    "@types/jest": "^26.0.10",
     "glob": "^7.1.6",
     "nodemon": "^2.0.2",
-    "pm2": "^4.4.0",
+    "pm2": "^4.4.1",
     "rimraf": "^3.0.1",
     "ts-json-schema-generator": "^0.71.1",
     "typescript": "^3.9.7"
