+++ services/auth/package.json
@@ -13,7 +13,7 @@
   },
   "devDependencies": {
     "@types/formidable": "^1.0.31",
-    "@types/node": "^14.0.27"
+    "@types/node": "^14.6.0"
   },
   "dependencies": {
     "@common/config": "1.0.0",
