+++ services/dashboard/package.json
@@ -12,9 +12,10 @@
     "build": "tsc -b"
   },
   "devDependencies": {
-    "@types/node": "^14.0.27"
+    "@types/node": "^14.6.0"
   },
   "dependencies": {
+    "@common/config": "1.0.0",
     "@common/service-utils": "1.0.0",
     "@furystack/auth-google": "^4.0.45",
     "@furystack/core": "^9.1.6",
