+++ services/diag/package.json
@@ -12,7 +12,7 @@
     "build": "tsc -b"
   },
   "devDependencies": {
-    "@types/node": "^14.0.27"
+    "@types/node": "^14.6.0"
   },
   "dependencies": {
     "@common/service-utils": "1.0.0",
