+++ services/media/package.json
@@ -13,7 +13,7 @@
   },
   "devDependencies": {
     "@types/amqplib": "^0.5.13",
-    "@types/node": "^14.0.27",
+    "@types/node": "^14.6.0",
     "@types/ws": "^7.2.5"
   },
   "dependencies": {
