+++ services/xpense/package.json
@@ -11,7 +11,7 @@
     "build": "tsc -b"
   },
   "devDependencies": {
-    "@types/node": "^14.0.27"
+    "@types/node": "^14.6.0"
   },
   "dependencies": {
     "@common/service-utils": "1.0.0",
