+++ workers/media-encoder/package.json
@@ -12,7 +12,7 @@
   },
   "devDependencies": {
     "@types/amqplib": "^0.5.13",
-    "@types/node": "^14.0.27",
+    "@types/node": "^14.6.0",
     "@types/rimraf": "^3.0.0"
   },
   "dependencies": {
